pyproject.toml
[project]
name = "huashengai-desktop"
version = "0.1.0"
description = "PyWebView desktop shell with a Vue frontend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "peewee>=3.17.9,<4",
    "playwright>=1.55,<2",
    "pywebview>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
